Fix oversight with how GL_EXT_texture_filter_anisotropic works
This fixes the issue where textures with NEAREST_LINEAR would come out with super awesome trileaner sampling when  gl_textureanisotropy had a value.

This also takes a look at the max AF value and sets the cvar to half of that upon startup for a cleaner look for first starting the game.